fix_handle_card-3
