+++ makefile
@@ -3,9 +3,12 @@ CC = gcc
 # LIBS =  -lm
 all: run
 
-run: server.c client.c uno.o structData.o
+run: uno.o structData.o interface.o server.c client.c
 	$(CC) uno.o structData.o server.c -o server
-	$(CC) `pkg-config --cflags gtk+-3.0` uno.o structData.o client.c -o client `pkg-config --libs gtk+-3.0` -export-dynamic -lpthread
+	$(CC) uno.o structData.o interface.o client.c -o client `pkg-config --cflags gtk+-3.0` `pkg-config --libs gtk+-3.0` -export-dynamic -lpthread
+
+interface.o : interface.c
+	$(CC) ${CFLAGS} interface.c `pkg-config --cflags gtk+-3.0` `pkg-config --libs gtk+-3.0` -export-dynamic -lpthread
 
 uno.o : uno.c
 	${CC} ${CFLAGS} uno.c
